chore: rewrite examples/nwc scripts in TypeScript
