+++ examples/nwc/client/get-balance.ts
@@ -4,28 +4,21 @@ import * as readline from "node:readline/promises";
 import { stdin as input, stdout as output } from "node:process";
 
 import { nwc } from "../../../dist/index.module.js";
+import type { nwc as NWC } from "../../../dist/index"; 
 
 const rl = readline.createInterface({ input, output });
 
-const nwcUrl =
-  process.env.NWC_URL ||
-  (await rl.question("Nostr Wallet Connect URL (nostr+walletconnect://...): "));
-
-const amount =
-  parseInt((await rl.question("Amount in sats (default 1 sat): ")) || "1") *
-  1000;
-
+const nwcUrl: string = process.env.NWC_URL || await rl.question("Nostr Wallet Connect URL (nostr+walletconnect://...): ");
 rl.close();
 
+// Initialize NWC client
 const client = new nwc.NWCClient({
   nostrWalletConnectUrl: nwcUrl,
-});
+}) as NWC.NWCClient;
 
-const response = await client.makeInvoice({
-  amount, // in millisats
-  description: "NWC Client example",
-});
-
+// Fetch balance
+const response = await client.getBalance();
 console.info(response);
 
+// Clean up
 client.close();
