+++ examples/nwc/client/get-info.ts
@@ -3,22 +3,22 @@ import "websocket-polyfill"; // required in node.js
 import * as readline from "node:readline/promises";
 import { stdin as input, stdout as output } from "node:process";
 
-import { webln as providers } from "../../dist/index.module.js";
+import { nwc } from "../../../dist/index.module.js";
+import type { nwc as NWC } from "../../../dist/index"; 
+
 
 const rl = readline.createInterface({ input, output });
 
 const nwcUrl =
   process.env.NWC_URL ||
   (await rl.question("Nostr Wallet Connect URL (nostr+walletconnect://...): "));
-const invoice = await rl.question("Lightning invoice: ");
 rl.close();
 
-const webln = new providers.NostrWebLNProvider({
+const client = new nwc.NWCClient({
   nostrWalletConnectUrl: nwcUrl,
-});
-await webln.enable();
-const response = await webln.sendPayment(invoice);
+}) as NWC.NWCClient;
+const response = await client.getInfo();
 
 console.info(response);
 
-webln.close();
+client.close();
