+++ examples/nwc/client/get-wallet-service-supported-methods.ts
@@ -4,6 +4,8 @@ import * as readline from "node:readline/promises";
 import { stdin as input, stdout as output } from "node:process";
 
 import { nwc } from "../../../dist/index.module.js";
+import type { nwc as NWC } from "../../../dist/index"; 
+
 
 const rl = readline.createInterface({ input, output });
 
@@ -14,9 +16,12 @@ rl.close();
 
 const client = new nwc.NWCClient({
   nostrWalletConnectUrl: nwcUrl,
-});
-const response = await client.getInfo();
+}) as NWC.NWCClient;
 
-console.info(response);
+// ! Method getWalletServiceSupportedMethods does not exist in the NWCClient class
+// TODO implement this method or remove this example
+// const response = await client.getWalletServiceSupportedMethods();
+
+// console.info(response);
 
 client.close();
