+++ examples/nwc/client/list-transactions.ts
@@ -3,7 +3,9 @@ import "websocket-polyfill"; // required in node.js
 import * as readline from "node:readline/promises";
 import { stdin as input, stdout as output } from "node:process";
 
-import { webln as providers } from "../../dist/index.module.js";
+import { nwc } from "../../../dist/index.module.js";
+import type { nwc as NWC } from "../../../dist/index"; 
+
 
 const rl = readline.createInterface({ input, output });
 
@@ -12,13 +14,12 @@ const nwcUrl =
   (await rl.question("Nostr Wallet Connect URL (nostr+walletconnect://...): "));
 rl.close();
 
-const webln = new providers.NostrWebLNProvider({
+const client = new nwc.NWCClient({
   nostrWalletConnectUrl: nwcUrl,
-});
-await webln.enable();
+}) as NWC.NWCClient;
 
 const ONE_WEEK_IN_SECONDS = 60 * 60 * 24 * 7;
-const response = await webln.listTransactions({
+const response = await client.listTransactions({
   from: Math.floor(new Date().getTime() / 1000 - ONE_WEEK_IN_SECONDS),
   until: Math.ceil(new Date().getTime() / 1000),
   limit: 30,
@@ -33,4 +34,4 @@ console.info(
   response,
 );
 
-webln.close();
+client.close();
