+++ examples/nwc/client/lookup-invoice.ts
@@ -4,32 +4,32 @@ import * as readline from "node:readline/promises";
 import { stdin as input, stdout as output } from "node:process";
 
 import { nwc } from "../../../dist/index.module.js";
+import type { nwc as NWC } from "../../../dist/index"; 
+
 
 const rl = readline.createInterface({ input, output });
 
-const nwcUrl =
+const nwcUrl: string =
   process.env.NWC_URL ||
   (await rl.question("Nostr Wallet Connect URL (nostr+walletconnect://...): "));
+
+const invoiceOrPaymentHash = await rl.question("Invoice or payment hash: ");
 rl.close();
 
 const client = new nwc.NWCClient({
   nostrWalletConnectUrl: nwcUrl,
+}) as NWC.NWCClient;
+
+const response = await client.lookupInvoice({
+  // provide one of the below
+  invoice: invoiceOrPaymentHash.startsWith("ln")
+    ? invoiceOrPaymentHash
+    : undefined,
+  payment_hash: !invoiceOrPaymentHash.startsWith("ln")
+    ? invoiceOrPaymentHash
+    : undefined,
 });
 
-const ONE_WEEK_IN_SECONDS = 60 * 60 * 24 * 7;
-const response = await client.listTransactions({
-  from: Math.floor(new Date().getTime() / 1000 - ONE_WEEK_IN_SECONDS),
-  until: Math.ceil(new Date().getTime() / 1000),
-  limit: 30,
-  // type: "incoming",
-  // unpaid: true,
-});
-
-console.info(
-  response.transactions.length + " transactions, ",
-  response.transactions.filter((t) => t.type === "incoming").length +
-    " incoming",
-  response,
-);
+console.info(response);
 
 client.close();
