+++ examples/nwc/client/make-invoice.ts
@@ -3,24 +3,30 @@ import "websocket-polyfill"; // required in node.js
 import * as readline from "node:readline/promises";
 import { stdin as input, stdout as output } from "node:process";
 
-import { webln as providers } from "../../dist/index.module.js";
+import { nwc } from "../../../dist/index.module.js";
+import type { nwc as NWC } from "../../../dist/index";
 
 const rl = readline.createInterface({ input, output });
 
 const nwcUrl =
   process.env.NWC_URL ||
   (await rl.question("Nostr Wallet Connect URL (nostr+walletconnect://...): "));
+
+const amount =
+  parseInt((await rl.question("Amount in sats (default 1 sat): ")) || "1") *
+  1000;
+
 rl.close();
 
-const webln = new providers.NostrWebLNProvider({
+const client = new nwc.NWCClient({
   nostrWalletConnectUrl: nwcUrl,
-});
-await webln.enable();
-const response = await webln.makeInvoice({
-  amount: 1, // in sats
-  defaultMemo: "NWC WebLN example",
+}) as NWC.NWCClient;
+
+const response = await client.makeInvoice({
+  amount, // in millisats
+  description: "NWC Client example",
 });
 
 console.info(response);
 
-webln.close();
+client.close();
