+++ examples/nwc/client/sign-message.ts
@@ -4,29 +4,23 @@ import * as readline from "node:readline/promises";
 import { stdin as input, stdout as output } from "node:process";
 
 import { nwc } from "../../../dist/index.module.js";
+import type { nwc as NWC } from "../../../dist/index"; 
 
 const rl = readline.createInterface({ input, output });
 
 const nwcUrl =
   process.env.NWC_URL ||
   (await rl.question("Nostr Wallet Connect URL (nostr+walletconnect://...): "));
+
 rl.close();
 
 const client = new nwc.NWCClient({
   nostrWalletConnectUrl: nwcUrl,
+}) as NWC.NWCClient;
+const response = await client.signMessage({
+  message: "Hello, world!",
 });
 
-const onNotification = (notification) =>
-  console.info("Got notification", notification);
+console.info(response);
 
-const unsub = await client.subscribeNotifications(onNotification);
-
-console.info("Waiting for notifications...");
-process.on("SIGINT", function () {
-  console.info("Caught interrupt signal");
-
-  unsub();
-  client.close();
-
-  process.exit();
-});
+client.close();
