+++ examples/nwc/client/subscribe.ts
@@ -4,6 +4,7 @@ import * as readline from "node:readline/promises";
 import { stdin as input, stdout as output } from "node:process";
 
 import { nwc } from "../../../dist/index.module.js";
+import type { nwc as NWC } from "../../../dist/index"; // Importing types for better type checking
 
 const rl = readline.createInterface({ input, output });
 
@@ -14,9 +15,19 @@ rl.close();
 
 const client = new nwc.NWCClient({
   nostrWalletConnectUrl: nwcUrl,
+}) as NWC.NWCClient;
+
+const onNotification = (notification) =>
+  console.info("Got notification", notification);
+
+const unsub = await client.subscribeNotifications(onNotification);
+
+console.info("Waiting for notifications...");
+process.on("SIGINT", function () {
+  console.info("Caught interrupt signal");
+
+  unsub();
+  client.close();
+
+  process.exit();
 });
-const response = await client.getWalletServiceSupportedMethods();
-
-console.info(response);
-
-client.close();
