+++ examples/nwc/get-info.ts
@@ -4,20 +4,23 @@ import * as readline from "node:readline/promises";
 import { stdin as input, stdout as output } from "node:process";
 
 import { webln as providers } from "../../dist/index.module.js";
+import type { webln as WebLN } from "../../dist/index"
+
 
 const rl = readline.createInterface({ input, output });
 
 const nwcUrl =
   process.env.NWC_URL ||
   (await rl.question("Nostr Wallet Connect URL (nostr+walletconnect://...): "));
-
 rl.close();
 
 const webln = new providers.NostrWebLNProvider({
   nostrWalletConnectUrl: nwcUrl,
-});
+}) as unknown as WebLN.NostrWebLNProvider;
+
+
 await webln.enable();
-const response = await webln.signMessage("Hello, world!");
+const response = await webln.getInfo();
 
 console.info(response);
