+++ examples/nwc/keysend.ts
@@ -4,6 +4,7 @@ import * as readline from "node:readline/promises";
 import { stdin as input, stdout as output } from "node:process";
 
 import { webln as providers } from "../../dist/index.module.js";
+import type { webln as WebLN } from "../../dist/index"
 
 const rl = readline.createInterface({ input, output });
 
@@ -11,21 +12,22 @@ const nwcUrl =
   process.env.NWC_URL ||
   (await rl.question("Nostr Wallet Connect URL (nostr+walletconnect://...): "));
 
-const invoiceOrPaymentHash = await rl.question("Invoice or payment hash: ");
 rl.close();
 
 const webln = new providers.NostrWebLNProvider({
   nostrWalletConnectUrl: nwcUrl,
-});
+}) as unknown as WebLN.NostrWebLNProvider;
+
+
 await webln.enable();
-const response = await webln.lookupInvoice({
-  // provide one of the below
-  paymentRequest: invoiceOrPaymentHash.startsWith("ln")
-    ? invoiceOrPaymentHash
-    : undefined,
-  paymentHash: !invoiceOrPaymentHash.startsWith("ln")
-    ? invoiceOrPaymentHash
-    : undefined,
+const response = await webln.keysend({
+  amount: 1,
+  destination:
+    "030a58b8653d32b99200a2334cfe913e51dc7d155aa0116c176657a4f1722677a3",
+  customRecords: {
+    696969: "017rsl75kNnSke4mMHYE", // hello@getalby.com
+    34349334: "example keysend message",
+  },
 });
 
 console.info(response);
