+++ examples/nwc/list-transactions.ts
@@ -4,29 +4,36 @@ import * as readline from "node:readline/promises";
 import { stdin as input, stdout as output } from "node:process";
 
 import { webln as providers } from "../../dist/index.module.js";
+import type { webln as WebLN } from "../../dist/index"
 
 const rl = readline.createInterface({ input, output });
 
 const nwcUrl =
   process.env.NWC_URL ||
   (await rl.question("Nostr Wallet Connect URL (nostr+walletconnect://...): "));
-
 rl.close();
 
 const webln = new providers.NostrWebLNProvider({
   nostrWalletConnectUrl: nwcUrl,
-});
+}) as unknown as WebLN.NostrWebLNProvider;
+
+
 await webln.enable();
-const response = await webln.keysend({
-  amount: 1,
-  destination:
-    "030a58b8653d32b99200a2334cfe913e51dc7d155aa0116c176657a4f1722677a3",
-  customRecords: {
-    696969: "017rsl75kNnSke4mMHYE", // hello@getalby.com
-    34349334: "example keysend message",
-  },
+
+const ONE_WEEK_IN_SECONDS = 60 * 60 * 24 * 7;
+const response = await webln.listTransactions({
+  from: Math.floor(new Date().getTime() / 1000 - ONE_WEEK_IN_SECONDS),
+  until: Math.ceil(new Date().getTime() / 1000),
+  limit: 30,
+  // type: "incoming",
+  // unpaid: true,
 });
 
-console.info(response);
+console.info(
+  response.transactions.length + " transactions, ",
+  response.transactions.filter((t) => t.type === "incoming").length +
+    " incoming",
+  response,
+);
 
 webln.close();
