+++ examples/nwc/make-invoice.ts
@@ -3,7 +3,8 @@ import "websocket-polyfill"; // required in node.js
 import * as readline from "node:readline/promises";
 import { stdin as input, stdout as output } from "node:process";
 
-import { nwc } from "../../../dist/index.module.js";
+import { webln as providers } from "../../dist/index.module.js";
+import type { webln as WebLN } from "../../dist/index";
 
 const rl = readline.createInterface({ input, output });
 
@@ -12,11 +13,17 @@ const nwcUrl =
   (await rl.question("Nostr Wallet Connect URL (nostr+walletconnect://...): "));
 rl.close();
 
-const client = new nwc.NWCClient({
+const webln = new providers.NostrWebLNProvider({
   nostrWalletConnectUrl: nwcUrl,
+}) as unknown as WebLN.NostrWebLNProvider;
+
+
+await webln.enable();
+const response = await webln.makeInvoice({
+  amount: 1, // in sats
+  defaultMemo: "NWC WebLN example",
 });
-const response = await client.getBalance();
 
 console.info(response);
 
-client.close();
+webln.close();
