+++ examples/nwc/send-multi-payment.ts
@@ -6,6 +6,7 @@ import * as readline from "node:readline/promises";
 import { stdin as input, stdout as output } from "node:process";
 
 import { webln as providers } from "../../dist/index.module.js";
+import type { webln as WebLN } from "../../dist/index";
 
 const rl = readline.createInterface({ input, output });
 
@@ -34,7 +35,8 @@ rl.close();
 
 const webln = new providers.NostrWebLNProvider({
   nostrWalletConnectUrl: nwcUrl,
-});
+}) as unknown as WebLN.NostrWebLNProvider;
+
 await webln.enable();
 try {
   const response = await webln.sendMultiPayment(invoices);
